Clean-up statistics in dataset

+++ org.eclipse.january.test/src/org/eclipse/january/dataset/IntegerDatasetTest.java
@@ -48,8 +48,8 @@ public class IntegerDatasetTest {
 		assertEquals(11., a.max().doubleValue(), 1e-6);
 		assertEquals(0., a.min().doubleValue(), 1e-6);
 		assertEquals(5.5, ((Number) a.mean()).doubleValue(), 1e-6);
-		assertEquals(3.6055512754639891, a.stdDeviation().doubleValue(), 1e-6);
-		assertEquals(13., a.variance().doubleValue(), 1e-6);
+		assertEquals(3.6055512754639891, a.stdDeviation(), 1e-6);
+		assertEquals(13., a.variance(), 1e-6);
 	}
 	
 	@Test
